requirements.txt
Flask==0.12.3
gunicorn==19.9.0
psycopg2-binary==2.7.6.1
